consolidate tc-ui <-> demo interface into single file
