--- .snyk
@@ -1,9 +1,0 @@
-# Snyk (https://snyk.io) policy file, patches or ignores known vulnerabilities.
-version: v1.13.5
-# ignores vulnerabilities until expiry date; change duration by modifying expiry date
-ignore:
-  'snyk:lic:npm:strong-cluster-control:Artistic-2.0':
-    - '@financial-times/n-cluster > strong-cluster-control':
-        reason: None given
-        expires: '2019-12-04T16:57:02.128Z'
-patch: {}
